// libs/admin-plugins/shipping/international-shipping/src/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
  Date: any;
};
// ...
export enum DeletionResult {
  /** The entity was successfully deleted */
  Deleted = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NotDeleted = 'NOT_DELETED'
}
// ...
export type DeletionResponse = {
  __typename?: 'DeletionResponse';
  result: DeletionResult;
  message?: Maybe<Scalars['String']>;
};

export type ConfigArgInput = {
  name: Scalars['String'];
  value: Scalars['String'];
};

export type ConfigurableOperationInput = {
  code: Scalars['String'];
  arguments: Array<ConfigArgInput>;
};

export type PaginatedList = {
  items: Array<Node>;
  totalItems: Scalars['Int'];
};

export type Node = {
  id: Scalars['ID'];
};

export enum SortOrder {
  Asc = 'ASC',
  Desc = 'DESC'
}
// ...
export type StringOperators = {
  eq?: Maybe<Scalars['String']>;
  contains?: Maybe<Scalars['String']>;
  in?: Maybe<Array<Scalars['String']>>;
  regex?: Maybe<Scalars['String']>;
};

export type BooleanOperators = {
  eq?: Maybe<Scalars['Boolean']>;
};

export type NumberRange = {
  start: Scalars['Float'];
  end: Scalars['Float'];
};

export type NumberOperators = {
  eq?: Maybe<Scalars['Float']>;
  lt?: Maybe<Scalars['Float']>;
  lte?: Maybe<Scalars['Float']>;
  gt?: Maybe<Scalars['Float']>;
  gte?: Maybe<Scalars['Float']>;
  between?: Maybe<NumberRange>;
};
// ...
export type ShippingCountryPrice = Node & {
  __typename?: 'ShippingCountryPrice';
  id: Scalars['ID'];
  country_code: Scalars['String'];
  country_name: Scalars['String'];
  price: Scalars['Float'];
};

export type ShippingCountryPriceList = PaginatedList & {
  __typename?: 'ShippingCountryPriceList';
  items: Array<ShippingCountryPrice>;
  totalItems: Scalars['Int'];
};

export type CreateShippingCountryInput = {
  country_code: Scalars['String'];
  price: Scalars['Float'];
};

export type UpdateShippingCountryInput = {
  id: Scalars['ID'];
  price: Scalars['Float'];
};
// ...
export type ShippingCountryPriceListOptions = {
  skip?: Maybe<Scalars['Int']>;
  take?: Maybe<Scalars['Int']>;
  sort?: Maybe<ShippingCountryPriceSortParameter>;
  filter?: Maybe<ShippingCountryPriceFilterParameter>;
};

export type ShippingCountryPriceSortParameter = {
  id?: Maybe<SortOrder>;
  country_code?: Maybe<SortOrder>;
  country_name?: Maybe<SortOrder>;
  price?: Maybe<SortOrder>;
};

export type ShippingCountryPriceFilterParameter = {
  country_code?: Maybe<StringOperators>;
  country_name?: Maybe<StringOperators>;
  price?: Maybe<NumberOperators>;
};